add basic url routing and introduce pages
app.vue becomes a layout component with the top nav menu
