+++ src/pages/AboutPage.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <h1>About</h1>
+    </div>
+</template>
